+++ app/api/api.go
@@ -701,7 +701,7 @@ func (a *api) start() error {
 			Token:        cfg.SRT.Token,
 			Logger:       a.log.logger.core.WithComponent("SRT").WithField("address", cfg.SRT.Address),
 			Collector:    a.sessions.Collector("srt"),
-			SRTLogTopics: []string{"handshake", "connection"},
+			SRTLogTopics: []string{"listen", "handshake", "connection"},
 		}
 
 		srtserver, err := srt.New(config)